tgbot/requirements.txt
psycopg2-binary
python-telegram-bot
